Rename the folder in types to be for table correctness related types
