+++ src/components/GuessTableCell/type.ts
@@ -5,7 +5,7 @@ import {
   TextDouble,
   TextImageCombo,
   TextSingle,
-} from "../../types/data/cellVariations.type";
+} from "../../types/tableCorrectness/cellVariations.type";
 
 type GuessTableCellProps = {
   cellData:
