+++ src/types/tableCorrectness/answerOptions.type.ts
