+++ src/types/tableCorrectness/cellVariations.type.ts
