+++ src/types/tableCorrectness/characterAnswerAccuracy.type.ts
