+++ src/types/tableCorrectness/characterTransformedData.type.ts
