+++ src/types/tableCorrectness/foodAnswerAccuracy.type.ts
