+++ src/types/tableCorrectness/weaponAnswerAccuracy.type.ts
