+++ src/types/tableCorrectness/weaponTransformedData.type.ts
